requirements.txt
streamlit
torch
diffusers
transformers
accelerate
safetensors
ffmpeg-python
Pillow
numpy
tqdm